+++ test/audits/forms/handler.test.js
@@ -20,11 +20,14 @@ import {
   formsAuditRunner,
   processOpportunityStep,
   runAuditAndSendUrlsForScrapingStep,
+  createAccessibilityOpportunity,
+  sendA11yUrlsForScrapingStep,
 } from '../../../src/forms-opportunities/handler.js';
 import { MockContextBuilder } from '../../shared.js';
 import formVitalsData from '../../fixtures/forms/formvitalsdata.json' with { type: 'json' };
 import expectedFormVitalsData from '../../fixtures/forms/expectedformvitalsdata.json' with { type: 'json' };
 import expectedFormSendToScraperData from '../../fixtures/forms/expectedformsendtoscraperdata.json' with { type: 'json' };
+import expectedFormA11yScraperData from '../../fixtures/forms/expectedforma11ysendtoscraperdata.json' with { type: 'json' };
 import { FORM_OPPORTUNITY_TYPES } from '../../../src/forms-opportunities/constants.js';
 
 use(sinonChai);
@@ -121,6 +124,139 @@ describe('audit and send scraping step', () => {
       granularity: 'hourly',
     });
     expect(result).to.deep.equal(expectedFormSendToScraperData);
+  });
+
+  it('send alteast 10 urls for scraping step if possible', async () => {
+    const formVitals = {
+      'form-vitals': [
+        {
+          url: 'https://example.com/form1',
+          formsubmit: {},
+          formview: { 'desktop:windows': 7898 },
+          formengagement: { 'desktop:windows': 100 },
+          pageview: { 'desktop:windows': 7898 },
+        },
+        {
+          url: 'https://example.com/form2', formsubmit: {}, formview: {}, formengagement: {}, pageview: {},
+        },
+        {
+          url: 'https://example.com/form3', formsubmit: {}, formview: {}, formengagement: {}, pageview: {},
+        },
+        {
+          url: 'https://example.com/form4', formsubmit: {}, formview: {}, formengagement: {}, pageview: {},
+        },
+        {
+          url: 'https://example.com/form5', formsubmit: {}, formview: {}, formengagement: {}, pageview: {},
+        },
+        {
+          url: 'https://example.com/form6', formsubmit: {}, formview: {}, formengagement: {}, pageview: {},
+        },
+        {
+          url: 'https://example.com/form7', formsubmit: {}, formview: {}, formengagement: {}, pageview: {},
+        },
+        {
+          url: 'https://example.com/form8', formsubmit: {}, formview: {}, formengagement: {}, pageview: {},
+        },
+        {
+          url: 'https://example.com/form9', formsubmit: {}, formview: {}, formengagement: {}, pageview: {},
+        },
+        {
+          url: 'https://example.com/form10', formsubmit: {}, formview: {}, formengagement: {}, pageview: {},
+        },
+        {
+          url: 'https://example.com/form11', formsubmit: {}, formview: {}, formengagement: {}, pageview: {},
+        },
+        {
+          url: 'https://example.com/form12', formsubmit: {}, formview: {}, formengagement: {}, pageview: {},
+        },
+        {
+          url: 'https://example.com/form13', formsubmit: {}, formview: {}, formengagement: {}, pageview: {},
+        },
+        {
+          url: 'https://example.com/form14', formsubmit: {}, formview: {}, formengagement: {}, pageview: {},
+        },
+      ],
+    };
+    context.rumApiClient.queryMulti = sinon.stub().resolves(formVitals);
+    const result = await runAuditAndSendUrlsForScrapingStep(context);
+    expect(result.urls.length).to.equal(10);
+  });
+});
+
+describe('send a11y urls for scraping step', () => {
+  let context;
+  const siteId = 'test-site-id';
+
+  // eslint-disable-next-line prefer-const
+  context = new MockContextBuilder()
+    .withSandbox(sandbox)
+    .withOverrides({
+      runtime: { name: 'aws-lambda', region: 'us-east-1' },
+      func: { package: 'spacecat-services', version: 'ci', name: 'test' },
+      site: {
+        getId: sinon.stub().returns(siteId),
+        getBaseURL: sinon.stub().returns('https://example.com'),
+        getLatestAuditByAuditType: sinon.stub().resolves({
+          auditResult: {
+            formVitals: formVitalsData['form-vitals'],
+            auditContext: {
+              interval: 15,
+            },
+          },
+          fullAuditRef: 'www.example.com',
+          siteId: 'test-site-id',
+        }),
+      },
+      env: {
+        S3_SCRAPER_BUCKET_NAME: 'test-bucket',
+      },
+      s3Client: {
+        send: sandbox.stub(),
+      },
+    })
+    .build();
+
+  beforeEach(() => {
+    // Mock the getScrapedDataForSiteId function response
+    context.s3Client.send.onCall(0).resolves({
+      Contents: [
+        { Key: 'scrapes/site-id/forms/scrape.json' },
+      ],
+      IsTruncated: false,
+    });
+
+    const mockFormResponseData = {
+      ContentType: 'application/json',
+      Body: {
+        transformToString: sandbox.stub().resolves(JSON.stringify({
+          finalUrl: 'https://www.business.adobe.com/newsletter',
+          scrapeResult: [
+            {
+              id: 'form1',
+              formType: 'newsletter',
+              visibleATF: true,
+              fieldCount: 3,
+              formSource: '#container-1 form.newsletter',
+            },
+          ],
+        })),
+      },
+    };
+
+    context.s3Client.send.resolves(mockFormResponseData);
+  });
+
+  afterEach(() => {
+    nock.cleanAll();
+    sinon.restore();
+  });
+
+  it('send a11y urls for scraping step', async () => {
+    const result = await sendA11yUrlsForScrapingStep(context);
+    expect(context.site.getLatestAuditByAuditType).calledWith('forms-opportunities');
+    // Verify that the s3Client was called to get the scraped data
+    expect(context.s3Client.send).to.have.been.called;
+    expect(result).to.deep.equal(expectedFormA11yScraperData);
   });
 });
 
@@ -252,5 +388,80 @@ describe('process opportunity step', () => {
     expect(result).to.deep.equal({
       status: 'complete',
     });
+  });
+});
+
+describe('send a11y issues to mystique', () => {
+  let context;
+  const siteId = 'test-site-id';
+
+  beforeEach(() => {
+    context = new MockContextBuilder()
+      .withSandbox(sandbox)
+      .withOverrides({
+        runtime: { name: 'aws-lambda', region: 'us-east-1' },
+        func: { package: 'spacecat-services', version: 'ci', name: 'test' },
+        site: {
+          getId: sinon.stub().returns(siteId),
+        },
+        log: {
+          info: sinon.stub(),
+        },
+      })
+      .build();
+
+    context.dataAccess.Opportunity.allBySiteIdAndStatus = sandbox.stub().resolves([]);
+    context.dataAccess.Opportunity.create = sandbox.stub().resolves();
+  });
+
+  afterEach(() => {
+    nock.cleanAll();
+    sinon.restore();
+  });
+
+  it('should not create opportunities when no a11y data is present', async () => {
+    const latestAudit = {
+      siteId: 'test-site-id',
+      auditId: 'test-audit-id',
+      getSiteId: () => 'test-site-id',
+      getAuditId: () => 'test-audit-id',
+    };
+
+    const scrapedData = {
+      formA11yData: [],
+    };
+
+    await createAccessibilityOpportunity(latestAudit, scrapedData, context);
+    expect(context.log.info).to.have.been.calledWith('[Form Opportunity] [Site Id: test-site-id] No a11y data found');
+  });
+
+  it('should create opportunities when a11y issues are present', async () => {
+    const latestAudit = {
+      siteId: 'test-site-id',
+      auditId: 'test-audit-id',
+      getSiteId: () => 'test-site-id',
+      getAuditId: () => 'test-audit-id',
+    };
+
+    const scrapedData = {
+      formA11yData: [{
+        a11yResult: [{
+          finalUrl: 'https://example.com/form1',
+          formSource: '#form1',
+          a11yIssues: [{
+            issue: 'Missing alt text',
+            level: 'error',
+            successCriterias: ['1.1.1'],
+            htmlWithIssues: '<img src="test.jpg">',
+            recommendation: 'Add alt text to image',
+          }],
+        }],
+      }],
+    };
+
+    await createAccessibilityOpportunity(latestAudit, scrapedData, context);
+
+    expect(context.dataAccess.Opportunity.create).to.have.been.called;
+    expect(context.log.info).to.have.been.calledWith('[Form Opportunity] [Site Id: test-site-id] a11y issues created');
   });
 });